My take on PID
So general idea goes like this. Function initialzePID should be called once in setup to specify user parameters. Then in loop function PID will be called with either 1 or 2 parameters. User can either call it with requested value and current value or with his own error. This comes from the idea of block representation of PID control, where in flows the error e and out goes output u (as it is for didactic purposes). More explanation can be found in the code. Unfortunately, I have no idea how will the code behave as I have nothing to test it on.