my version of the sampling function
I deleted the whole part of the branch. This branch was created for finalizing the sampling function, 
and just contains my version of the function. After assembling and finalizing the final function can be put into the 
AutomationShield class or into another one. 
The only problem is that I can not hide the interrupt service routine (yet). 
I think the solution might be using the same method which was also used by Richard.