Pulled greybox identification from master
I have pulled the greybox identification example from master into FloatShield branch to be able to work on it, and make changes. Removed the two identification examples that are already incorporated in the FloatShield_Ident_Greybox.m. Changed the output unit from m to mm. Changed the radius of the ball to 0.015 [m] (15mm), what caused lower fits in the nonlinear model  (66.24%) , will try to get it back up somehow.